Move JobInfoCard to own component file
